etl/requirements.txt
pandas
psycopg2-binary
pyodbc
python-dotenv
APScheduler
